Moved tuple_of_11 to separate file
